requirements.txt
Babel==2.6.0
certifi==2018.4.16
chardet==3.0.4
Django==1.11.14
django-channels==0.7.0
django-phonenumber-field==2.0.0
django-rest-auth==0.9.3
djangorestframework==3.8.2
idna==2.7
oauthlib==2.1.0
phonenumberslite==8.9.10
PyJWT==1.6.4
python-openid==2.2.5
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
rest-social-auth==1.3.1
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
typing==3.6.4
urllib3==1.23
